Add arc-swap dependency and refactor indexer for improved concurrency
- Introduced `arc-swap` crate to enable lock-free snapshots of indexes.
- Updated `Indexer` struct to use `ArcSwap` for index storage, enhancing performance.
- Refactored methods in `Indexer` to utilize the new `ArcSwap` functionality.
- Removed obsolete fork lock methods to streamline index management during forks.
- Adjusted various functions to ensure compatibility with the new index handling.

+++ graph/src/runtime/functions/algo_procedures.rs
@@ -2214,7 +2214,7 @@ fn register_harmonic_centrality(funcs: &mut Functions) {
 
                 runtime.check_timeout()?;
                 let mut nodes: GrB_Vector = null_mut();
-                let node_vec_len = compact_to_id.as_ref().map_or(g.node_count() as u64, |m| m.len() as u64);
+                let node_vec_len = compact_to_id.as_ref().map_or(g.node_count(), |m| m.len() as u64);
                 GrB_Vector_new(&raw mut nodes, GrB_BOOL, node_vec_len);
                 GrB_Vector_assign_BOOL(
                     nodes,
